requirements.txt
hvicorn
pyyaml